+++ src/Loader/Preprocessors/Handlebars/helpers/block/block.js
@@ -21,14 +21,8 @@ module.exports = (Handlebars) => {
    * @return {string}
    */
   return function (name, options) {
-    const context = this;
-    let partial = context._blocks[name] || options.fn;
+    const partial = this._blocks[name] || options.fn;
 
-    if (typeof partial === 'string') {
-      partial = Handlebars.compile(partial);
-      context._blocks[name] = partial;
-    }
-
-    return partial(context, { data: options.hash });
+    return partial(this, { data: options.hash });
   };
 };