+++ test/cases/integrity-dynamic-chunks/expected/src_chunk_js.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([["src_chunk_js"],{457:()=>{console.log(">> chunk file")}}]);