+++ test/cases/integrity-split-chunks/src/js/app.js
@@ -1,0 +1,2 @@
+import mainA from './main-a';
+console.log('>> app: ', { mainA });