+++ test/cases/integrity-split-chunks/src/js/lib.js
@@ -1,0 +1,16 @@
+const Lib = {
+  methodA() {
+    // 166 chars
+    return '### Module AAA: Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam volu';
+  },
+
+  methodB() {
+    return '### Module BBB: Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam volu';
+  },
+
+  methodC() {
+    return '*** Common used module: Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam volu';
+  },
+};
+
+export { Lib };